+++ src/lib/domain/repositories/init_repository.dart
@@ -5,5 +5,6 @@ abstract class InitRepository {
   Future<void> setInitialized();
 
   Future<void> initHomePage(BuildContext context);
-  Future<void> initSmartlists(BuildContext context);
+  Future<void> createFavoritesSmartlist(BuildContext context);
+  Future<void> createNewlyAddedSmartlist(BuildContext context);
 }