+++ src/lib/main.dart
@@ -123,7 +123,6 @@ class _RootPageState extends State<RootPage> {
         );
 
         initRepository.initHomePage(context);
-        initRepository.initSmartlists(context);
       }
     });